requirements.txt
appnope==0.1.0
awsebcli==3.10.1
blessed==1.14.2
botocore==1.5.38
cement==2.8.2
colorama==0.3.7
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.6
django-cors-headers==2.0.2
django-extensions==1.7.7
django-filter==1.0.2
django-rest-auth==0.9.1
djangorestframework==3.6.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
ipython==5.3.0
ipython-genutils==0.2.0
jmespath==0.9.2
oauthlib==2.0.2
olefile==0.44
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
python-dateutil==2.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
selenium==3.3.3
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.10.0
social-auth-core==1.2.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
websocket-client==0.40.0
